app/ai_microservice/requirements.txt
fastapi
uvicorn
faiss-cpu
openai
PyPDF2
numpy
python-multipart